pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Farid Delgado <farid.delgado@thomsonreuters.com>"]

[tool.poetry.dependencies]
python = "~3.7"
google-cloud-vision = "^2.6.2"
Flask = "^2.0.2"
matplotlib = "^3.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
